requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain>=0.2.0
langchain-anthropic>=0.1.0
anthropic>=0.25.0

# Kubernetes
kubernetes>=28.0.0

# Vector DB
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.25.0
tenacity>=8.2.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
